+++ app/src/main/java/org/simple/clinic/deniedaccess/AccessDeniedModel.kt
@@ -1,0 +1,7 @@
+package org.simple.clinic.deniedaccess
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+class AccessDeniedModel : Parcelable