+++ app/src/main/java/org/simple/clinic/deniedaccess/AccessDeniedScreenKey.kt
@@ -1,16 +1,14 @@
 package org.simple.clinic.deniedaccess
 
-import android.os.Parcelable
 import kotlinx.android.parcel.IgnoredOnParcel
 import kotlinx.android.parcel.Parcelize
-import org.simple.clinic.R
-import org.simple.clinic.router.screen.FullScreenKey
+import org.simple.clinic.navigation.v2.ScreenKey
 
 @Parcelize
-data class AccessDeniedScreenKey(val fullName: String) : FullScreenKey, Parcelable {
+data class AccessDeniedScreenKey(val fullName: String) : ScreenKey() {
 
   @IgnoredOnParcel
   override val analyticsName = "Access Denied"
 
-  override fun layoutRes() = R.layout.screen_access_denied
+  override fun instantiateFragment() = AccessDeniedScreen()
 }